Move ProcessWithFeedback into ProcessRunner
